# docs/api.yaml
openapi: 3.0.3
info:
  title: WhatToEat API
  version: 1.0.0
  description: |
    今天吃什么（WhatToEat）后端 API 文档。

    本文档覆盖当前已经实现的资源：认证、餐厅查询、公开评论、AI 推荐、黑名单 CRUD 与备注 CRUD。
servers:
  - url: http://127.0.0.1:8080
    description: WeChat DevTools local development
  - url: http://{host}:8080
    description: Real-device LAN development (replace with your host machine IP)
    variables:
      host:
        default: 192.168.0.10
        description: Host machine LAN IP for WeChat real-device debugging
...
paths:
  /api/v1/auth/wechat-login:
    post:
      tags: [Auth]
      summary: 微信登录
      operationId: wechatLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatLoginRequest'
      responses:
        '201':
          description: 登录成功并创建会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 登出
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 获取当前登录用户
      operationId: me
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/users/{userId}/choice-history:
    post:
      tags: [UserSignals]
      summary: 记录用户最近吃过的餐厅
      operationId: createChoiceHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChoiceHistoryRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [UserSignals]
      summary: 分页查询最近选择历史
      operationId: listChoiceHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoiceHistoryPageSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/users/{userId}/restaurant-reviews/{poiId}:
    get:
      tags: [Reviews]
      summary: 查询当前用户对某餐厅的评论
      operationId: getRestaurantReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RestaurantPoiId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDetailSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ReviewNotFound'
    put:
      tags: [Reviews]
      summary: 创建或更新当前用户对某餐厅的评论
      operationId: upsertRestaurantReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RestaurantPoiId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRestaurantReviewRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDetailSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/AiOrAmapUpstreamError'
        '504':
          $ref: '#/components/responses/AiOrAmapUpstreamTimeout'
    delete:
      tags: [Reviews]
      summary: 删除当前用户对某餐厅的评论
      operationId: deleteRestaurantReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RestaurantPoiId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ReviewNotFound'
        '502':
          $ref: '#/components/responses/AiOrAmapUpstreamError'
        '504':
          $ref: '#/components/responses/AiOrAmapUpstreamTimeout'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: "自定义 session token，通过 Authorization: Bearer <token> 传递，不是 JWT。"
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: 必填；用户主键 ID。请先调用 /api/v1/auth/me 获取当前登录用户的 id，再将该值填入路径中；不能留空。若使用 Apifox，可将环境变量 {{userId}} 绑定到该路径参数。
      example: 1
...
    RestaurantPoiId:
      name: poiId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: .*(\S|[^\s]).*
      description: 餐厅的高德 POI ID。
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 可选；页码，从 1 开始，默认 1。
      example: 1
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: 可选；每页条数，默认 10，最大 100。
      example: 10
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationFailed:
              value:
                code: 1001
                message: 参数校验失败
                data: null
            noteContentInvalid:
              value:
                code: 2003
                message: 备注内容非法
                data: null
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: 1003
                message: 未登录或token无效
                data: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            userNotFound:
              value:
                code: 1002
                message: 用户不存在
                data: null
            noteNotFound:
              value:
                code: 2004
                message: 备注不存在
                data: null
...
    ReviewNotFound:
      description: 评论不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            reviewNotFound:
              value:
                code: 2104
                message: 评论不存在
                data: null
    AiOrAmapUpstreamError:
      description: AI 或高德上游失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            amapError:
              value:
                code: 3001
                message: 高德上游失败
                data: null
            aiError:
              value:
                code: 3004
                message: AI 服务失败
                data: null
    AiOrAmapUpstreamTimeout:
      description: AI 或高德上游超时
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            amapTimeout:
              value:
                code: 3002
                message: 高德上游超时
                data: null
            aiTimeout:
              value:
                code: 3005
                message: AI 服务超时
                data: null
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
      required: [code, message, data]
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true
          example: null
      required: [code, message, data]
    WechatLoginRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          description: 小程序端 wx.login() 返回的 code；开发环境如启用 mock 登录，也可传 mock-code-*。
          pattern: .*(\S|[^\s]).*
          example: mock-code-001
        nickname:
          type: string
          description: 昵称按纯文本保存；后端会移除 HTML / script 等标签。
          example: Alice
        avatarUrl:
          type: string
          nullable: true
          example: https://placehold.co/200x200/F97316/FFFFFF?text=%E5%90%83
      required: [code]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: 8f2c5f7314d649d3a9124f1930c68dfd
        user:
          $ref: '#/components/schemas/UserInfo'
      required: [token, user]
    LoginSuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/LoginResponse'
      required: [code, message, data]
    UserInfoSuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/UserInfo'
      required: [code, message, data]
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        openid:
          type: string
          example: mock-openid-mock-code-001
        nickname:
          type: string
          example: Alice
        avatarUrl:
          type: string
          nullable: true
          example: https://placehold.co/200x200/F97316/FFFFFF?text=%E5%90%83
      required: [id, openid]
...
    CreateChoiceHistoryRequest:
      type: object
      properties:
        poiId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: .*(\S|[^\s]).*
          example: B0FFCH001
        poiName:
          type: string
          maxLength: 128
          nullable: true
          example: 兰州拉面
      required: [poiId]
    ChoiceHistoryItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        poiId:
          type: string
          example: B0FFCH001
        poiName:
          type: string
          nullable: true
          example: 兰州拉面
        chosenAt:
          type: string
          format: date-time
          example: '2026-04-18T12:00:00'
      required: [id, poiId, chosenAt]
    ChoiceHistoryPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChoiceHistoryItem'
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        total:
          type: integer
          format: int64
          example: 3
      required: [items, page, size, total]
    ChoiceHistoryPageSuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/ChoiceHistoryPageResponse'
      required: [code, message, data]
...
    UpsertRestaurantReviewRequest:
      type: object
      properties:
        poiNameSnapshot:
          type: string
          maxLength: 128
          nullable: true
          example: 兰州拉面
        ratingScore:
          type: number
          format: double
          description: 必填；支持 0.5 步进，取值范围 0.5 到 5.0。
          example: 4.5
        perCapitaPrice:
          type: integer
          description: 必填；人均价格，单位元，必须大于 0。
          example: 29
        content:
          type: string
          description: 必填；评论内容会按纯文本保存，移除 HTML / script 标签后去除首尾空白，净化后不能为空，长度不能超过 1000。
          maxLength: 1000
          example: 汤底稳，性价比不错
      required: [ratingScore, perCapitaPrice, content]
    ReviewDetail:
      type: object
      properties:
        poiId:
          type: string
          example: B0FFREVIEW001
        poiNameSnapshot:
          type: string
          nullable: true
          example: 沙县小吃
        ratingScore:
          type: number
          format: double
          example: 4.5
        perCapitaPrice:
          type: integer
          example: 28
        content:
          type: string
          example: 出餐快
        createdAt:
          type: string
          format: date-time
          example: '2026-04-17T12:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-17T12:30:00'
      required: [poiId, ratingScore, perCapitaPrice, content, createdAt, updatedAt]
...
    ReviewDetailSuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/ReviewDetail'
      required: [code, message, data]
...